Docker/requirements_docker.txt
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.4.2
tensorflow-cpu==2.15.0
joblib